+++ alea/model.py
@@ -3,6 +3,7 @@ import warnings
 from pydoc import locate
 from copy import deepcopy
 from typing import List, Tuple, Callable, Optional, Union
+from itertools import product
 
 import numpy as np
 from scipy.optimize import brentq
@@ -11,7 +12,7 @@ from blueice.likelihood import _needs_data
 from inference_interface import toydata_to_file
 
 from alea.parameters import Parameters
-from alea.utils import within_limits, clip_limits, asymptotic_critical_value
+from alea.utils import within_limits, clip_limits, asymptotic_critical_value, ReadOnlyDict
 
 
 class StatisticalModel:
@@ -217,7 +218,9 @@ class StatisticalModel:
             metadata (dict, optional (default=None)): metadata to store with the data.
                 If None, no metadata is stored.
         """
-        if all([isinstance(d, dict) for d in data_list]):
+        if all([isinstance(d, dict) for d in data_list]) or all(
+            [isinstance(d, ReadOnlyDict) for d in data_list]
+        ):
             _data_list = [list(d.values()) for d in data_list]
         elif all([isinstance(d, list) for d in data_list]):
             _data_list = data_list
@@ -304,7 +307,9 @@ class StatisticalModel:
         return cost
 
     @_needs_data
-    def fit(self, verbose=False, **kwargs) -> Tuple[dict, float]:
+    def fit(
+        self, verbose=False, disable_index_fitting=False, max_index_fitting_iter=10, **kwargs
+    ) -> Tuple[dict, float]:
         """Fit the model to the data by maximizing the likelihood. Return a dict containing best-fit
         values of each parameter, and the value of the likelihood evaluated there. While the
         optimization is a minimization, the likelihood returned is the __maximum__ of the
@@ -312,6 +317,9 @@ class StatisticalModel:
 
         Args:
             verbose (bool): if True, print the Minuit object
+            disable_index_fitting (bool): if True, disable the index fitting
+                even if the model has index parameters.
+            max_index_fitting_iter (int): maximum number of iterations for index fitting
 
         Returns:
             dict, float: best-fit values of each parameter,
@@ -335,13 +343,67 @@ class StatisticalModel:
         for par in fixed_params:
             m.fixed[par] = True
 
-        # Call migrad to do the actual minimization
-        m.migrad()
+        # Get the index parameters, which could have problem if simply using migrad
+        index_parameters = [
+            p for p in self.parameters if p.ptype == "index" and p.name not in fixed_params
+        ]
+
+        if disable_index_fitting or (len(index_parameters) == 0):
+            # Call migrad to do the actual minimization
+            m = self._migrad_fit(m)
+        else:
+            m = self._migrad_index_mixing_fit(m, index_parameters, max_index_fitting_iter, verbose)
+
         self.minuit_object = m
         if verbose:
             print(m)
         # alert! This gives the _maximum_ likelihood
         return m.values.to_dict(), -1 * m.fval
+
+    def _migrad_fit(self, m):
+        m.migrad()
+        return m
+
+    def _migrad_index_mixing_fit(self, m, index_parameters, max_index_fitting_iter, verbose):
+        index_anchors = [p.blueice_anchors for p in index_parameters]
+        index_names = [p.name for p in index_parameters]
+        index_grid = [
+            {index_names[i]: anchor[i] for i in range(len(anchor))}
+            for anchor in product(*index_anchors)
+        ]
+
+        # We fix the index parameters in migrad
+        for par in index_names:
+            m.fixed[par] = True
+
+        # We firstly do optimization on other parameters with index parameters
+        # fixed to their initial guesses. Then we grid search over the index
+        # parameters given the optimized parameters. We repeat the optimization
+        # and grid search until the optimization converges
+        for itr in range(max_index_fitting_iter):
+            m.migrad()
+
+            # Find the best-fit index parameters
+            lls = np.zeros(len(index_grid))
+            for i in range(len(lls)):
+                best_fit_params = m.values.to_dict()
+                best_fit_params.update(index_grid[i])
+                lls[i] = self.ll(**best_fit_params)
+            for var in index_names:
+                m.values[var] = index_grid[np.argmax(lls)][var]
+
+            # Calculating Hessian will update the validity of
+            # the fitting given the new index parameters
+            m.hesse()
+            if m.valid:
+                break
+
+        if verbose and itr == max_index_fitting_iter - 1:
+            print(
+                "The index searching iteration times reached the maximum! "
+                "The optimization could not converge!"
+            )
+        return m
 
     def _confidence_interval_checks(
         self,