+++ alea/utils.py
@@ -4,6 +4,7 @@ import json
 import yaml
 import importlib_resources
 import itertools
+import blueice
 from glob import glob
 from copy import deepcopy
 from pydoc import locate
@@ -12,8 +13,11 @@ from hashlib import sha256
 from base64 import b32encode
 from collections.abc import Mapping
 from typing import Any, List, Dict, Tuple, Optional, Union, cast, get_args, get_origin
+from blueice.pdf_morphers import Morpher
+from itertools import product
 
 import h5py
+import matplotlib.pyplot as plt
 
 # These imports are needed to evaluate strings
 import numpy  # noqa: F401
@@ -55,6 +59,15 @@ class ReadOnlyDict:
         raise TypeError(
             "This dictionary is read-only, please initialize a new one in order to change it."
         )
+
+    def keys(self):
+        return self._data.keys()
+
+    def values(self):
+        return self._data.values()
+
+    def items(self):
+        return self._data.items()
 
 
 def evaluate_numpy_scipy_expression(value: str):
@@ -104,6 +117,26 @@ def get_analysis_space(analysis_space: list) -> list:
     return eval_analysis_space
 
 
+def _prefix_file_path(
+    config: dict, template_folder_list: list, ignore_keys: List[str] = ["name", "histname"]
+):
+    """Prefix file path with template_folder_list whenever possible.
+
+    Args:
+        config (dict): dictionary contains file path
+        template_folder_list (list): list of possible base folders. Ordered by priority.
+        ignore_keys (list, optional (default=["name", "histname"])):
+        keys to be ignored when prefixing
+
+    """
+    for key in config.keys():
+        if isinstance(config[key], str) and key not in ignore_keys:
+            try:
+                config[key] = get_file_path(config[key], template_folder_list)
+            except RuntimeError:
+                pass
+
+
 def adapt_likelihood_config_for_blueice(
     likelihood_config: dict, template_folder_list: list
 ) -> dict:
@@ -123,6 +156,8 @@ def adapt_likelihood_config_for_blueice(
     likelihood_config_copy["analysis_space"] = get_analysis_space(
         likelihood_config_copy["analysis_space"]
     )
+
+    _prefix_file_path(likelihood_config_copy, template_folder_list)
 
     if "default_source_class" in likelihood_config_copy:
         default_source_class = locate(likelihood_config_copy["default_source_class"])
@@ -145,6 +180,7 @@ def adapt_likelihood_config_for_blueice(
                 get_file_path(template_filename, template_folder_list)
                 for template_filename in source["template_filenames"]
             ]
+        _prefix_file_path(source, template_folder_list)
     return likelihood_config_copy
 
 
@@ -286,7 +322,7 @@ def get_template_folder_list(likelihood_config, extra_template_path: Optional[st
     # Add extra_template_path to the end of template_folder_list
     if extra_template_path is not None:
         template_folder_list.append(extra_template_path)
-    return template_folder_list
+    return list(set(template_folder_list))
 
 
 def asymptotic_critical_value(
@@ -608,3 +644,69 @@ def deterministic_hash(thing, length=10):
     # disable bandit
     digest = sha256(jsonned.encode("ascii")).digest()
     return b32encode(digest)[:length].decode("ascii").lower()
+
+
+def signal_multiplier_estimator(
+    signal: np.ndarray,
+    background: np.ndarray,
+    data: np.ndarray,
+    iteration=100,
+    diagnostic=False,
+) -> float:
+    """Estimate the best-fit signal multiplier using perturbation theory. The method tries to solve
+    the critial point of the likelihood function by perturbation theory, where the likelihood
+    function is defined as the binned Poisson likelihood function, given signal, background models
+    and data.
+
+    Args:
+        signal (np.ndarray): signal model
+        background (np.ndarray): background model
+        data (np.ndarray): data array
+        iteration (int, optional (default=100)): number of iterations
+    Returns:
+        float: best-fit signal multiplier
+
+    """
+    mask = (signal > 0) | (background > 0)
+    if np.any(data[~mask] > 0):
+        raise ValueError("Data has non-zero values where signal and background is zero.")
+
+    sig = signal[mask].ravel()
+    bkg = background[mask].ravel()
+    obs = data[mask].ravel()
+
+    @np.errstate(invalid="ignore", divide="ignore")
+    def correction_on_multiplier(x):
+        exp = sig * x + bkg
+        return np.sum(np.where(exp > 0, (obs / exp - 1) * sig, 0)) / np.sum(
+            np.where(exp > 0, obs * sig**2 / exp**2, 0)
+        )
+
+    # For underfluctutation case, the best-fit multiplier could be negative
+    # in which case the perturbation theory may not converge or be negative.
+    # Thus we clip it to be non-negative.
+    x = np.sum(obs - bkg) / np.sum(sig)
+    xs = [x]
+    for _ in range(iteration):
+        x += correction_on_multiplier(x)
+        x = np.clip(x, 0, None)
+        xs.append(x)
+    if diagnostic:
+        plt.plot(xs, marker=".")
+        plt.xlabel("Iteration")
+        plt.ylabel("x")
+    return x
+
+
+class IndexMorpher(Morpher):
+    """IndexMorpher is a morpher which applies no interpolation."""
+
+    def get_anchor_points(self, bounds, n_models=None):
+        grid = [par.keys() for _, (par, _, _) in self.shape_parameters.items()]
+        return list(product(*grid))
+
+    def make_interpolator(self, f, extra_dims, anchor_models):
+        return lambda z: f(anchor_models[tuple(z)])
+
+
+blueice.pdf_morphers.MORPHERS["IndexMorpher"] = IndexMorpher